+++ Konsole.Tests/BufferedWriterTests/BufferTests.cs
@@ -10,7 +10,7 @@ namespace Konsole.Tests.BufferedWriterTests
             [Test]
             public void not_return_lines_not_written_to()
             {
-                var console = new BufferedWriter(10, 10);
+                var console = new Window(10, 10, false);
                 console.WriteLine("1");
                 console.WriteLine("2");
                 var lines = console.BufferWritten;
@@ -24,7 +24,7 @@ namespace Konsole.Tests.BufferedWriterTests
             [Test]
             public void return_all_lines_in_between_any_lines_written_to()
             {
-                var console = new BufferedWriter(10, 10);
+                var console = new Window(10, 10, false);
                 console.PrintAt(1, 1, "A");
                 console.PrintAt(3, 3, "B");
                 var lines = console.BufferWritten;
@@ -44,7 +44,7 @@ namespace Konsole.Tests.BufferedWriterTests
             [Test]
             public void return_all_lines()
             {
-                var con = new BufferedWriter(10, 2);
+                var con = new Window(10, 2, false);
                 con.WriteLine("one");
                 con.WriteLine("two");
                 Assert.AreEqual(new[] { "one       ", "two       " }, con.Buffer);
@@ -53,7 +53,7 @@ namespace Konsole.Tests.BufferedWriterTests
             [Test]
             public void not_be_trimmed()
             {
-                var con = new BufferedWriter(10, 2);
+                var con = new Window(10, 2, false);
                 Assert.AreEqual(new[] { "          ", "          " }, con.Buffer);
             }