+++ Konsole.Tests/BufferedWriterTests/CursorTests.cs
@@ -12,7 +12,7 @@ namespace Konsole.Tests.BufferedWriterTests
             [Test]
             public void CursorLeft_should_return_the_x_position_that_the_next_character_will_be_written_to()
             {
-                var console = new BufferedWriter(30, 20);
+                var console = new Window(30, 20, false);
                 Assert.AreEqual(0, console.CursorLeft);
                 console.Write("Today ");
                 Assert.AreEqual(6, console.CursorLeft);
@@ -27,10 +27,34 @@ namespace Konsole.Tests.BufferedWriterTests
                 console.Write("123");
                 Assert.AreEqual(1, console.CursorTop);
             }
+
+            [Test]
+            public void Setting_cursor_position_should_set_cursor_x_and_y_position()
+            {
+                // #ADH : while this tests passes, it doesnt prove that echo causes the correct out to be displayed. Something to think about! mmm...interesting.
+                var console = new Window(10, 20, false);
+                console.CursorLeft = 5;
+                console.CursorTop = 1;
+                console.Write("4");
+                Assert.AreEqual(new[] { "          ","     4    " }, console.BufferWritten);
+            }
+
+            [Test]
+            public void Setting_cursor_left_should_set_cursor_x_position()
+            {
+                var console = new Window(10, 20, false);
+                console.Write("123");
+                Assert.AreEqual(3,console.CursorLeft);
+                Assert.AreEqual(0, console.CursorTop);
+                console.CursorLeft = 5;
+                console.Write("4");
+                Assert.AreEqual(new[] {"123  4    "}, console.BufferWritten);
+            }
+
             [Test]
             public void setting_CursorLeft_position_should_change_x_position_without_affecting_y_position_and_allow_writing_at_different_x_positions()
             {
-                var console = new BufferedWriter(80, 20);
+                var console = new Window(80, 20, false);
                 Assert.AreEqual(0, console.CursorTop);
                 console.WriteLine("line1");
                 Assert.AreEqual(1, console.CursorTop);
@@ -50,7 +74,7 @@ namespace Konsole.Tests.BufferedWriterTests
             [Test]
             public void calling_writeline_should_increment_cursortop_position()
             {
-                var console = new BufferedWriter(80, 20);
+                var console = new Window(80, 20, false);
                 Assert.AreEqual(0, console.CursorTop);
                 console.WriteLine("line1");
                 Assert.AreEqual(1, console.CursorTop);
@@ -67,7 +91,7 @@ namespace Konsole.Tests.BufferedWriterTests
             [Test]
             public void setting_cursor_top_to_a_previously_written_line_should_allow_us_to_overwrite_previously_written_lines()
             {
-                var console = new BufferedWriter(80, 20);
+                var console = new Window(80, 20, false);
                 console.WriteLine("line 0");
                 console.WriteLine("line 1");
                 console.WriteLine("line 2");
@@ -88,7 +112,7 @@ namespace Konsole.Tests.BufferedWriterTests
         [Test]
         public void PrintAt_tests()
         {
-            var console = new BufferedWriter(5, 5);
+            var console = new Window(5, 5, false);
             console.PrintAt(0, 0, "*");
             console.PrintAt(2, 2, "*");
             console.PrintAt(4, 4, "*");
@@ -114,8 +138,28 @@ namespace Konsole.Tests.BufferedWriterTests
             System.Console.WriteLine(console.BufferWrittenString);
             Assert.That(console.BufferWrittenTrimmed, Is.EqualTo(trimmed));
             Assert.That(console.BufferWritten, Is.EqualTo(buffer));
-
         }
+
+
+        [Test]
+        public void PrintAt_with_text_causing_overflow_should_overflow_to_next_line()
+        {
+            var console = new Window(5, 5, false);
+            console.PrintAt(2, 2, "12345");
+
+            var expected = new[]
+            {
+                "*    ",
+                "     ",
+                "  123",
+                "45   ",
+                "     "
+            };
+
+            System.Console.WriteLine(console.BufferWrittenString);
+            Assert.AreEqual(expected, console.Buffer);
+        }
+