+++ Konsole.Tests/Lines/LineTests.cs
@@ -19,7 +19,7 @@ namespace Konsole.Tests
         [Test]
         public void draw_box_should_draw_box()
         {
-            var console = new BufferedWriter(200, 20);
+            var console = new Window(200, 20, false);
             // draw box 40 wide, and 6 high
             new Draw(console).Box(2, 2, 42, 8, "my test box", LineThickNess.Single);
             Approvals.Verify(console.BufferWrittenString);
@@ -28,7 +28,7 @@ namespace Konsole.Tests
         [Test]
         public void should_be_able_to_draw_complex_forms_with_mixed_lines()
         {
-            var console = new BufferedWriter(200, 50);
+            var console = new Window(200, 50, false);
             int height = 18;
             int sy = 2;
             int sx = 2;
@@ -54,7 +54,7 @@ namespace Konsole.Tests
         [Test]
         public void should_support_drawing_any_positive_size_boxes()
         {
-            var console = new BufferedWriter(200, 100);
+            var console = new Window(200, 100, false);
             var line = new Draw(console);
 
             // negative width box should not render anything
@@ -98,7 +98,7 @@ namespace Konsole.Tests
         {
             using (ApprovalResults.ForScenario(firstThickness, secondThickness, merge))
             {
-                var console = new BufferedWriter(80, 35);
+                var console = new Window(80, 35, false);
                 console.WriteLine("box1 :{0}, box2:{1}, MergeOrOverlap:{2}", firstThickness, secondThickness, merge);
                 var line = new Draw(console, firstThickness, merge);