+++ change-log.md
@@ -28,6 +28,7 @@ New core `Window` functionality, plus big class renames, main interface `IConsol
 
 ### Backlog
 
+- fix bug with `PrintAt` with text that overflows, causing crash.
 - Optional borders for windows
 - Merge two window borders to create neat single border.
 - allow for printing off the screen without crashing.